% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{accessor_create}
\alias{accessor_create}
\title{Create accessor function for each dependency}
\usage{
accessor_create(.x, do_update)
}
\arguments{
\item{.x}{\code{(chr)} The full path of the file.}

\item{do_update}{\code{(lgl)} If TRUE, will update the dependency.}
}
\value{
A function that can be used to access the dependency.
}
\description{
Create accessor function for each dependency
}
