% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{beds_available_between}
\alias{beds_available_between}
\title{beds_available_between}
\usage{
beds_available_between(., start = ReportStart, end = ReportEnd)
}
\description{
beds_available_between
}
\seealso{
Other *_between: 
\code{\link{entered_between}()},
\code{\link{exited_between}()},
\code{\link{operating_between}()},
\code{\link{served_between}()},
\code{\link{stayed_between}()}
}
\concept{*_between}
