% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{between_df}
\alias{between_df}
\title{between_df}
\usage{
between_df(., status, start = ReportStart, end = ReportEnd)
}
\arguments{
\item{.}{\code{(data.frame/tibble)} Input to be filtered. In a `magrittr` pipe this will always be the first object}

\item{status}{\code{(character)} One of:
\itemize{
  \item{\code{"served"/"se"}}
  \item{\code{"stayed"/"st"}}
  \item{\code{"entered"/"en"}}
  \item{\code{"exited"/"ex"}}
}
that specifies the type of function to be performed}

\item{start}{The ReportStart variable created from user input - will be automatically retrieved from parent environments if not specified. If start is named other than ReportStart, it must be specified.}

\item{end}{The ReportEnd variable created from user input - will be automatically retrieved from parent environments if not specified. If end is named other than ReportEnd, it must be specified.}
}
\value{
\code{data.frame} after filtering the appropriate columns
}
\description{
Performs quick filtering of qpr_* data.frames with the input of the type of filtering
}
\keyword{Internal}
