% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{create_accessors}
\alias{create_accessors}
\title{Create accessor functions in the namespace (local files only)}
\usage{
create_accessors(
  dep_dir = "data",
  deps = NULL,
  dep_update = update_local,
  update_all = TRUE
)
}
\arguments{
\item{dep_dir}{\code{(chr)} The directory where dependencies are stored.}

\item{deps}{\code{(chr)} Full file paths for dependencies to be accessed.}

\item{dep_update}{\code{(function)} Dependency update function.}

\item{update_all}{\code{(lgl)} Whether to load/update all dependencies.}
}
\value{
Accessor functions will be available in the package namespace.
}
\description{
Create accessor functions in the namespace (local files only)
}
