% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{datatable_add_bars}
\alias{datatable_add_bars}
\title{Add \link[DT]{styleColorBar} or `styleDivergentBar` to datatable}
\usage{
datatable_add_bars(table, columns, valueColumns, ..., divergent = FALSE)
}
\arguments{
\item{table}{a table object created from \code{\link[DT]{datatable}()}}

\item{columns}{the indices of the columns to be formatted (can be character,
numeric, logical, or a formula of the form \code{~ V1 + V2}, which is
equivalent to \code{c('V1', 'V2')})}

\item{valueColumns}{indices of the columns from which the cell values are
obtained; this can be different with the \code{columns} argument, e.g. you
may style one column based on the values of a different column}

\item{...}{
  Arguments passed on to \code{\link[DT:formatCurrency]{DT::formatStyle}}, \code{\link[DT:styleInterval]{DT::styleColorBar}}, \code{\link[=styleDivergentBar]{styleDivergentBar}}
  \describe{
    \item{\code{target}}{the target to apply the CSS styles to (the current cell or the
full row)}
    \item{\code{fontWeight}}{the font weight, e.g. \code{'bold'} and \code{'normal'}}
    \item{\code{color}}{the font color, e.g. \code{'red'} and \code{'#ee00aa'}}
    \item{\code{backgroundColor}}{the background color of table cells}
    \item{\code{background}}{the background of table cells}
    \item{\code{data}}{a numeric vector whose range will be used for scaling the
table data from 0-100 before being represented as color bars. A vector
of length 2 is acceptable here for specifying a range possibly wider or
narrower than the range of the table data itself.}
    \item{\code{angle}}{a number of degrees representing the direction to fill the
gradient relative to a horizontal line and the gradient line, going
counter-clockwise. For example, 90 fills right to left and -90 fills
left to right.}
    \item{\code{color_pos}}{The color of the bars for the positive values}
    \item{\code{color_neg}}{The color of the bars for the negative values}
  }}

\item{divergent}{\code{(logical)} Whether to use `styleDivergentBar`}
}
\value{
\code{(datatable)}
}
\description{
Add \link[DT]{styleColorBar} or `styleDivergentBar` to datatable
}
