% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{datatable_options_update}
\alias{datatable_options_update}
\title{Update datatable options}
\usage{
datatable_options_update(x, options, hide_cols)
}
\arguments{
\item{x}{\code{(datatables)}}

\item{options}{\code{(list)} of options to replace}
}
\value{
\code{(datatables)}
}
\description{
Update datatable options
}
