% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{do_assignment}
\alias{do_assignment}
\title{Assign accessor functions to the package namespace}
\usage{
do_assignment(funs, ns = pkgload::pkg_name())
}
\arguments{
\item{funs}{A list of accessor functions to assign.}

\item{ns}{The package namespace where functions will be assigned.}
}
\value{
Functions are assigned to the namespace or returned as a list.
}
\description{
Assign accessor functions to the package namespace
}
