% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{entered_between}
\alias{entered_between}
\title{entered_between}
\usage{
entered_between(., start = ReportStart, end = ReportEnd)
}
\description{
entered_between
}
\seealso{
Other *_between: 
\code{\link{beds_available_between}()},
\code{\link{exited_between}()},
\code{\link{operating_between}()},
\code{\link{served_between}()},
\code{\link{stayed_between}()}
}
\concept{*_between}
