% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{exited_between}
\alias{exited_between}
\title{exited_between}
\usage{
exited_between(., start = ReportStart, end = ReportEnd)
}
\description{
exited_between
}
\seealso{
Other *_between: 
\code{\link{beds_available_between}()},
\code{\link{entered_between}()},
\code{\link{operating_between}()},
\code{\link{served_between}()},
\code{\link{stayed_between}()}
}
\concept{*_between}
