% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils_helpers.R
\name{find_path}
\alias{find_path}
\title{find_path}
\usage{
find_path(.fn, n = 4)
}
\arguments{
\item{.fn}{The case-sensitive regex for the file passed to \code{\link[base]{list.files}}.}

\item{n}{the number of parent directories to traverse before erroring}
}
\value{
The objects are loaded into the specified environment.
}
\description{
Finds a file or directory by traversing the directory tree.
}
\examples{
find_path("Rminor.RData")
}
