% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{iterate}
\alias{iterate}
\title{Iterative generate output functions}
\usage{
iterate(
  x,
  fn,
  outputId,
  env = rlang::caller_env(),
  ...,
  rc = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{x}{\code{(list)} of items to iterate over}

\item{fn}{\code{(fn)} output function to apply}

\item{outputId}{\code{(character)} The namespace ID (1,2,3 will be appended for each iteration)}

\item{...}{Further arguments passed on to `fn`}

\item{header_names}{\code{(logical)} Whether to create \link[shiny]{h4} headers above each item using the name}

\item{ns}{\code{(function)} ns function from the enclosing shiny context}
}
\value{
\code{(shiny.tag.list)}
}
\description{
Iterative generate output functions
}
