% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{make_columns}
\alias{make_columns}
\title{Make columns from assorted shiny.tag elements
Sorts shiny.tags into columns based on the maximum number of columns (`max_cols`) per row}
\usage{
make_columns(x, max_cols = TRUE, fn = list(bs4Dash::box, bs4Dash::column)[[1]])
}
\arguments{
\item{x}{\code{(shiny.tags)}}

\item{max_cols}{\code{(logical/integer)} Either `TRUE` **Default** for a default of 4 columns per row, `FALSE` for no columns, or an integer indicating the max number of columns.}
}
\value{
\code{(list(s))}
}
\description{
Make columns from assorted shiny.tag elements
Sorts shiny.tags into columns based on the maximum number of columns (`max_cols`) per row
}
