% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils_helpers.R
\name{qend_date}
\alias{qend_date}
\title{qend_date}
\usage{
qend_date(.)
}
\arguments{
\item{.}{\code{(character)} The input to transform}
}
\value{
\code{(Date)} The corresponding Date
}
\description{
Transform quarter slider text input to end date for filtering
}
\examples{
\dontrun{
ReportEnd <- qend_date(input$slider)
}
}
\keyword{Internal}
