% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_qpr_expr.R
\docType{data}
\name{qpr_expr}
\alias{qpr_expr}
\title{qpr_expr}
\format{
An object of class \code{list} of length 15.
}
\usage{
qpr_expr
}
\description{
List to hold all module dependency expressions
}
\keyword{datasets}
