% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{qpr_infobox}
\alias{qpr_infobox}
\title{qpr_infobox}
\usage{
qpr_infobox(
  .data,
  .replace = FALSE,
  title = "Average Score",
  color = "purple",
  value = .data$AvgScore,
  icon = shiny::icon("shopping-cart"),
  subtitle = "See table below for detail.",
  ...
)
}
\arguments{
\item{.data}{\code{(any)} Data to be passed in and used subsequent arguments}

\item{.replace}{\code{(logical)} whether to replace the default arguments with those supplied and eliminate the default arguments, or to replace existing defaults & and add additional args specified}

\item{title}{Info box title.}

\item{color}{A color for the box. Valid colors are defined as follows:
\itemize{
  \item \code{primary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#007bff")}.
  \item \code{secondary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6c757d")}.
  \item \code{info}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#17a2b8")}.
  \item \code{success}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#28a745")}.
  \item \code{warning}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ffc107")}.
  \item \code{danger}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#dc3545")}.
  \item \code{gray-dark}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#343a40")}.
  \item \code{gray}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#adb5bd")}.
  \item \code{white}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#fff")}.
  \item \code{indigo}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6610f2")}.
  \item \code{lightblue}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#3c8dbc")}.
  \item \code{navy}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#001f3f")}.
  \item \code{purple}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#605ca8")}.
  \item \code{fuchsia}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#f012be")}.
  \item \code{pink}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#e83e8c")}.
  \item \code{maroon}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#d81b60")}.
  \item \code{orange}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ff851b")}.
  \item \code{lime}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#01ff70")}.
  \item \code{teal}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#39cccc")}.
  \item \code{olive}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#3d9970")}.
}}

\item{value}{The value to display in the box. Usually a number or short text.}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{subtitle}{Any extra UI element.}

\item{...}{
  Arguments passed on to \code{\link[bs4Dash:infoBox]{bs4Dash::infoBox}}
  \describe{
    \item{\code{width}}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}
    \item{\code{href}}{An optional URL to link to.}
    \item{\code{fill}}{If FALSE (the default), use a white background for the content, and
the color argument for the background of the icon. If TRUE, use the color argument
for the background of the content; the icon will use the same color with a slightly
darkened background.}
    \item{\code{gradient}}{Whether to use gradient style for background color. Default to FALSE.}
    \item{\code{elevation}}{Box elevation.}
    \item{\code{iconElevation}}{Icon elevation compared to the main content (relief). 3 by default.}
    \item{\code{tabName}}{Optional: \link[bs4Dash]{infoBox} behaves like \link[bs4Dash]{menuItem} and
may be used to navigate between multiple \link[bs4Dash]{tabItem}.}
  }}
}
\description{
Function to render infobox from default template for QPR tabitems
}
