% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{qpr_plotly}
\alias{qpr_plotly}
\title{qpr_plotly}
\usage{
qpr_plotly(
  .d,
  title,
  x_col = "clients",
  xaxis_title = "",
  y_col = "Average",
  yaxis_title = "Measure",
  y_label = "Average Length of Stay",
  hover_text = NULL,
  mode = "markers",
  project_type = NULL,
  rect_above_line = TRUE,
  percent_format = FALSE,
  goals = list(),
  ...
)
}
\arguments{
\item{.d}{\code{(data.frame)} Input data}

\item{title}{\code{(character/list)}}

\item{...}{\code{(named lists)} of arguments passed on to \link[plotly]{layout}. These will take the place of boilerplate values, preserving unspecified boilerplate parameters. If boilerplate parameters should be replaced entirely with specified values, set `.sub = FALSE`. For boilerplate layout parameters, see examples.}

\item{x}{\code{(formula)} Formula for variable on x-axis in plotly tilde format}

\item{xaxis}{\code{(list)} of x-axis label parameters}

\item{y}{\code{(formula)} Formula for variable on y-axis in plotly tilde format}

\item{yaxis}{\code{(list)} of y-axis label parameters}

\item{.sub}{\code{(logical)} flag to specify whether boilerplate parameters should be switched out for those specified to `...` or replaced entirely.}
}
\value{
\code{(plotly)} graph object to be passed to \link[plotly]{renderPlotly}
}
\description{
A boilerplate plot_ly layout for QPR tab bar graphs with arguments for values that change across tabs.
}
