% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{qstart_date}
\alias{qstart_date}
\title{qstart_date}
\usage{
qstart_date(input)
}
\arguments{
\item{.}{\code{(character)} The input to transform}
}
\value{
\code{(Date)} The corresponding Date
}
\description{
Transform quarter slider text input to start date for filtering
}
\examples{
\dontrun{
ReportStart <- qstart_date(input$slider)
}
}
\keyword{Internal}
