% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{served_between}
\alias{served_between}
\title{served_between}
\usage{
served_between(., start = ReportStart, end = ReportEnd)
}
\arguments{
\item{.}{\code{(data.frame/tibble)} Input to be filtered. In a `magrittr` pipe this will always be the first object}

\item{start}{The ReportStart variable created from user input - will be automatically retrieved from parent environments if not specified. If start is named other than ReportStart, it must be specified.}

\item{end}{The ReportEnd variable created from user input - will be automatically retrieved from parent environments if not specified. If end is named other than ReportEnd, it must be specified.}
}
\value{
\code{data.frame} after filtering the appropriate columns
}
\description{
Filters a data.frame between the start and end dates specified
}
\examples{
\dontrun{
ReportStart = Sys.Date() - lubridate::weeks(4)
ReportEnd = Sys.Date()
qpr_leavers \%>\% served_between()
}
}
\seealso{
Other *_between: 
\code{\link{beds_available_between}()},
\code{\link{entered_between}()},
\code{\link{exited_between}()},
\code{\link{operating_between}()},
\code{\link{stayed_between}()}
}
\concept{*_between}
