% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{server_header}
\alias{server_header}
\title{Create a default header block}
\usage{
server_header(title, ..., program, date_range, region, county)
}
\arguments{
\item{title}{\code{(character)} Title to be wrapped in \link[shiny]{h2}}

\item{program}{\code{(character)} Program name to be wrapped in \link[shiny]{h4}}

\item{date_range}{\code{(Date)} vector to be collapsed with `" - "` in \link[shiny]{h4}}
}
\value{
\code{shiny.tag.list}
}
\description{
Create a default header block
}
