% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{stayed_between}
\alias{stayed_between}
\title{stayed_between}
\usage{
stayed_between(., start = ReportStart, end = ReportEnd)
}
\description{
stayed_between
}
\seealso{
Other *_between: 
\code{\link{beds_available_between}()},
\code{\link{entered_between}()},
\code{\link{exited_between}()},
\code{\link{operating_between}()},
\code{\link{served_between}()}
}
\concept{*_between}
