% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_QPR_tabItem.R
\docType{data}
\name{tab_choices}
\alias{tab_choices}
\title{Tab Choices}
\format{
An object of class \code{list} of length 4.
}
\usage{
tab_choices
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{program_choices}{\code{(named list)} of choices of program types for the radio picker. Default does not show the UI item - choices must be provided for the picker to show.}

\item{region_choices}{\code{(named list)} of choices of regions for the region drop-down selector. Default does not show the UI item - choices must be provided for the picker to show options}

\item{radio_mean}{\code{(logical)} whether to show the Mean/Median based average radio UI}
}
\description{
A shiny Module to generate the QPR tabitems.
}
\details{
QPR_tabItem UI Function
}
\keyword{datasets}
