% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{ui_date_range}
\alias{ui_date_range}
\title{A date range picker with sensible defaults}
\usage{
ui_date_range(
  inputId = rlang::caller_env()$ns("date_range"),
  label = "Date Range",
  start = prev_month(),
  end = Sys.Date(),
  min = rm_dates()$meta_HUDCSV$Export_Start,
  width = 300,
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{start}{The initial start date. Either a Date object, or a string in
\code{yyyy-mm-dd} format. If NULL (the default), will use the current
date in the client's time zone.}

\item{end}{The initial end date. Either a Date object, or a string in
\code{yyyy-mm-dd} format. If NULL (the default), will use the current
date in the client's time zone.}

\item{min}{The minimum allowed date. Either a Date object, or a string in
\code{yyyy-mm-dd} format.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:reexports]{validateCssUnit()}}.}

\item{...}{
  Arguments passed on to \code{\link[shiny:dateRangeInput]{shiny::dateRangeInput}}
  \describe{
    \item{\code{max}}{The maximum allowed date. Either a Date object, or a string in
\code{yyyy-mm-dd} format.}
    \item{\code{format}}{The format of the date to display in the browser. Defaults to
\code{"yyyy-mm-dd"}.}
    \item{\code{startview}}{The date range shown when the input object is first clicked.
Can be "month" (the default), "year", or "decade".}
    \item{\code{weekstart}}{Which day is the start of the week. Should be an integer
from 0 (Sunday) to 6 (Saturday).}
    \item{\code{language}}{The language used for month and day names. Default is "en".
Other valid values include "ar", "az", "bg", "bs", "ca", "cs", "cy", "da",
"de", "el", "en-AU", "en-GB", "eo", "es", "et", "eu", "fa", "fi", "fo",
"fr-CH", "fr", "gl", "he", "hr", "hu", "hy", "id", "is", "it-CH", "it",
"ja", "ka", "kh", "kk", "ko", "kr", "lt", "lv", "me", "mk", "mn", "ms",
"nb", "nl-BE", "nl", "no", "pl", "pt-BR", "pt", "ro", "rs-latin", "rs",
"ru", "sk", "sl", "sq", "sr-latin", "sr", "sv", "sw", "th", "tr", "uk",
"vi", "zh-CN", and "zh-TW".}
    \item{\code{separator}}{String to display between the start and end input boxes.}
    \item{\code{autoclose}}{Whether or not to close the datepicker immediately when a
date is selected.}
  }}
}
\description{
A date range picker with sensible defaults
}
