% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{ui_header_row}
\alias{ui_header_row}
\title{The UI Header output}
\usage{
ui_header_row(
  outputId = rlang::caller_env()$ns("header"),
  width = 12,
  headerBorder = FALSE
)
}
\arguments{
\item{outputId}{\code{(namespaced character)} A character vector wrapped in `ns` from the parent environment.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default card width of 6 occupies 1/2 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}

\item{headerBorder}{Whether to display a border between the header and body.
TRUE by default}
}
\value{
A fluidrow containing a minimizable box with the header
}
\description{
The UI Header output
}
