% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{ui_list}
\alias{ui_list}
\title{Construct a list from various elements}
\usage{
ui_list(x, ..., l_style = NULL, ordered = FALSE)
}
\arguments{
\item{x}{\code{(data.frame)} with a **Required** `text` column and **Optional** `style` & `icon` columns}

\item{...}{named elements with which to make a data.frame.}

\item{ordered}{\code{(logical)}  Whether the list should be ordered `<ol>`}
}
\value{
\code{(shiny.tag)}
}
\description{
The `icon` is placed before `text`. Any additional arguments will be added after `text`.
}
