% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{update_local}
\alias{update_local}
\title{Update a dependency locally (no remote)}
\usage{
update_local(filepaths)
}
\arguments{
\item{filepaths}{\code{(chr)} A list of file paths to check and load.}
}
\value{
Loaded dependencies.
}
\description{
Update a dependency locally (no remote)
}
