% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{which_cols}
\alias{which_cols}
\title{Give DT ready column numbers from names or numbers}
\usage{
which_cols(x, .data)
}
\arguments{
\item{x}{\code{(character/numeric)} Column numbers or names}
}
\description{
Give DT ready column numbers from names or numbers
}
