% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{operating_between}
\alias{operating_between}
\title{operating_between}
\usage{
operating_between(., start = ReportStart, end = ReportEnd)
}
\description{
operating_between
}
\seealso{
Other *_between: 
\code{\link{beds_available_between}()},
\code{\link{entered_between}()},
\code{\link{exited_between}()},
\code{\link{served_between}()},
\code{\link{stayed_between}()}
}
\concept{*_between}
