% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{qpr_plotly}
\alias{qpr_plotly}
\title{qpr_plotly}
\usage{
qpr_plotly(
  .d,
  title,
  x = ~FriendlyProjectName,
  xaxis = list(title = ""),
  y = ~Percent,
  yaxis = list(title = "Households", tickformat = "\%"),
  ...,
  .sub = TRUE
)
}
\arguments{
\item{.d}{\code{(data.frame)} Input data}

\item{title}{\code{(character/list)}}

\item{xaxis}{\code{(list)} of x-axis label parameters}

\item{yaxis}{\code{(list)} of y-axis label parameters}

\item{...}{\code{(named lists)} of arguments passed on to \link[plotly]{layout}. These will take the place of boilerplate values, preserving unspecified boilerplate parameters. If boilerplate parameters should be replaced entirely with specified values, set `.sub = FALSE`. For boilerplate layout parameters, see examples.}

\item{.sub}{\code{(logical)} flag to specify whether boilerplate parameters should be switched out for those specified to `...` or replaced entirely.}
}
\value{
\code{(plotly)} graph object to be passed to \link[plotly]{renderPlotly}
}
\description{
A boilerplate plot_ly layout for QPR tab bar graphs with arguments for values that change across tabs.
}
\examples{
\dontrun{
# Boilerplate plotly layout options
.opts_layout <- list(
p = .p,
xaxis = xaxis,
yaxis = yaxis,
title = list(
 text = title,
 font = list(
   size = 15
 )),
margin = list(
 l = 50,
 r = 50,
 b = 100,
 t = 100,
 pad = 4
),
shapes = list(
 type = "rect",
 name = "CoC Goal",
 fillcolor = "#008000",
 line = list(color = "white", width = .01),
 layer = "below",
 xref = "paper",
 yref = "y",
 x0 = 0,
 x1 = 1,
 y0 = ~ Goal[1],
 y1 = 1,
 opacity = .2
),
title = title
)
}
}
